--- gisprocessor/ImportToJson.py
@@ -1,9 +1,0 @@
-import json
-
-
-def read_json_file(file_path):
-    with open(file_path, 'r', encoding='utf-8') as file:
-        data = json.load(file)
-    lst = []
-    for key, value in data.items():
-        lst.append((key, value))
